refactor and fix error
