--- lib/screens/MarkAttendanceStudents.dart
@@ -1,16 +1,0 @@
-import 'constants.dart';
-import 'package:flutter/material.dart';
-
-class MarkAttendanceStudents extends StatefulWidget {
-  @override
-  _MarkAttendanceStudentsState createState() => _MarkAttendanceStudentsState();
-}
-
-class _MarkAttendanceStudentsState extends State<MarkAttendanceStudents> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body : Text("Yet to be implemented",)
-    );
-  }
-}
